Fix 'Removing Old Files'
Fixing the way deciding which file is removed.

+++ script/mysqldump.sh
@@ -27,7 +27,7 @@ function execute () {
 function dump_database() {
     DUMPED_TABLES=${TABLES[@]}
     COMMAND="$MYSQLDUMP $MYSQLDUMP_OPTIONS -u $DB_USERNAME --password=$DB_PASSWORD -h $DB_HOST $DB_NAME $DUMPED_TABLES"
-    eval "$COMMAND > $DUMP_FILENAME_FULL"
+    #eval "$COMMAND > $DUMP_FILENAME_FULL"
     
     compress_dump
     clearing_old_backup
@@ -35,7 +35,7 @@ function dump_database() {
 
 function compress_dump() {
     if [ $COMPRESSED_DUMP = 1 ]
-        then            
+        then
             COMPRESSED_FILENAME_FULL="$COMPRESSED_DIRECTORY$DUMP_FILENAME"
             
             # choosing the compress method
@@ -64,7 +64,7 @@ function clearing_dump() {
     if [ $REMOVE_DUMP_FILE = 1 ]
         then
             # only remove dump when compressed file is exist
-            if [ -e $DUMP_FILENAME_FULL ]
+            if [ -e $DUMP_FILENAME_FULL && ]
                 then
                     REMOVE_COMMAND="rm -rf $DUMP_FILENAME_FULL"
                     eval "$REMOVE_COMMAND"
@@ -75,14 +75,12 @@ function clearing_dump() {
 # Clears both compressed and sql dump file. Leaving between 
 # $KEEPSAKE_DAYS until today
 function clearing_old_backup() {
-    if [ $REMOVE_OLD_DUMP=1 ]
+    if [ $REMOVE_OLD_DUMP = 1 ]
         then
-            # 4 is reduction for each variable used in prefix
-            PREFIX_LENGTH=${#DB_HOST}+${#FILENAME_SEPARATOR}+${DB_NAME}+${FILENAME_SEPARATOR}
 
             # clearing dump (sql) file
             for FILE in $( ls ${DUMP_DIRECTORY} ); do
-                if [ "${FILE:$PREFIX_LENGTH:8}" -lt  "${KEEPSAKE_DAYS}" ]
+                if [[ ${FILE:$DUMP_FILENAME_DATE_POSITION:8} -lt  ${KEEPSAKE_DAYS} ]]
                     then
                         CLEAR_DUMP_COMMAND="rm -rf $DUMP_DIRECTORY$FILE"
                         eval "$CLEAR_DUMP_COMMAND"
@@ -91,8 +89,7 @@ function clearing_old_backup() {
 
             # clearing compressed file
             for FILE in $( ls ${COMPRESSED_DIRECTORY} ); do
-                #echo "$FILE :=: ${FILE:$PREFIX_LENGTH:8} :: ${KEEPSAKE_DAYS}"
-                if [[ ${FILE:$PREFIX_LENGTH:8} -lt  ${KEEPSAKE_DAYS} ]]
+                if [[ ${FILE:$DUMP_FILENAME_DATE_POSITION:8} -lt  ${KEEPSAKE_DAYS} ]]
                     then
                         CLEAR_COMPRESSED_COMMAND="rm -rf $COMPRESSED_DIRECTORY$FILE"
                         eval "$CLEAR_COMPRESSED_COMMAND"
